pkg/agent: move agent-related code from pkg/k8sutil
This code has nothing to do with Kubernetes, so it should live in
pkg/agent, not in pkg/k8sutil.
