+++ pkg/agent/agent_internal_test.go
@@ -1,4 +1,4 @@
-package k8sutil
+package agent
 
 import (
 	"reflect"
