+++ pkg/k8sutil/metadata.go
@@ -1,12 +1,8 @@
 package k8sutil
 
 import (
-	"bufio"
 	"context"
 	"fmt"
-	"io/ioutil"
-	"os"
-	"strings"
 
 	v1api "k8s.io/api/core/v1"
 	v1meta "k8s.io/apimachinery/pkg/apis/meta/v1"
@@ -15,13 +11,6 @@ import (
 	v1core "k8s.io/client-go/kubernetes/typed/core/v1"
 	watchtools "k8s.io/client-go/tools/watch"
 	"k8s.io/client-go/util/retry"
-	"k8s.io/klog/v2"
-)
-
-const (
-	updateConfPath         = "/usr/share/flatcar/update.conf"
-	updateConfOverridePath = "/etc/flatcar/update.conf"
-	osReleasePath          = "/etc/os-release"
 )
 
 // NodeAnnotationCondition returns a condition function that succeeds when a
@@ -158,90 +147,4 @@ func Unschedulable(nc v1core.NodeInterface, node string, sched bool) error {
 	}
 
 	return nil
-}
-
-// splitNewlineEnv splits newline-delimited KEY=VAL pairs and update map.
-func splitNewlineEnv(m map[string]string, envs string) {
-	sc := bufio.NewScanner(strings.NewReader(envs))
-	for sc.Scan() {
-		spl := strings.SplitN(sc.Text(), "=", 2)
-
-		// Just skip empty lines or lines without a value.
-		if len(spl) == 1 {
-			continue
-		}
-
-		m[spl[0]] = spl[1]
-	}
-}
-
-// VersionInfo contains Flatcar version and update information.
-type VersionInfo struct {
-	Name    string
-	ID      string
-	Group   string
-	Version string
-}
-
-func getUpdateMap() (map[string]string, error) {
-	infomap := map[string]string{}
-
-	// This file should always be present on Flatcar.
-	b, err := ioutil.ReadFile(updateConfPath)
-	if err != nil {
-		return nil, fmt.Errorf("reading file %q: %w", updateConfPath, err)
-	}
-
-	splitNewlineEnv(infomap, string(b))
-
-	updateConfOverride, err := ioutil.ReadFile(updateConfOverridePath)
-	if err != nil {
-		if !os.IsNotExist(err) {
-			return nil, fmt.Errorf("reading file %q: %w", updateConfOverridePath, err)
-		}
-
-		klog.Infof("Skipping missing update.conf: %w", err)
-	}
-
-	splitNewlineEnv(infomap, string(updateConfOverride))
-
-	return infomap, nil
-}
-
-func getReleaseMap() (map[string]string, error) {
-	infomap := map[string]string{}
-
-	// This file should always be present on Flatcar.
-	b, err := ioutil.ReadFile(osReleasePath)
-	if err != nil {
-		return nil, fmt.Errorf("reading file %q: %w", osReleasePath, err)
-	}
-
-	splitNewlineEnv(infomap, string(b))
-
-	return infomap, nil
-}
-
-// GetVersionInfo returns VersionInfo from the current Flatcar system.
-//
-// Should probably live in a different package.
-func GetVersionInfo() (*VersionInfo, error) {
-	updateconf, err := getUpdateMap()
-	if err != nil {
-		return nil, fmt.Errorf("unable to get update configuration: %w", err)
-	}
-
-	osrelease, err := getReleaseMap()
-	if err != nil {
-		return nil, fmt.Errorf("unable to get os release info: %w", err)
-	}
-
-	vi := &VersionInfo{
-		Name:    osrelease["NAME"],
-		ID:      osrelease["ID"],
-		Group:   updateconf["GROUP"],
-		Version: osrelease["VERSION"],
-	}
-
-	return vi, nil
 }
